+++ changelogs/custom/pypi/genshi.py
@@ -1,0 +1,7 @@
+"""Custom parser for Genshi"""
+
+URL = "https://genshi.edgewall.org/export/head/trunk/ChangeLog"
+
+
+def get_urls(releases, **kwargs):
+    return {URL}, set()