+++ tests/test_changelogs.py
@@ -18,6 +18,14 @@ def record(monkeypatch, betamax_session):
     monkeypatch.setattr("changelogs.changelogs.Session", session)
 
 
+def test_genshi():
+    # This also find URL in the description.
+    log = changelogs.get("genshi")
+    assert 'Add support for Python 3.1, 3.2 and 3.3' in log['0.7.0']
+    assert 'The builder API now accepts streams as children' in log['0.3.6']
+    assert 'project name was changed from "Markup" to "Genshi"' in log['0.3.0']
+
+
 def test_3to2():
     # This also find URL in the description.
     log = changelogs.get("3to2")