+++ tests/test_changelogs.py
@@ -18,6 +18,13 @@ def record(monkeypatch, betamax_session):
     monkeypatch.setattr("changelogs.changelogs.Session", session)
 
 
+def test_python_ldap():
+    log = changelogs.get("python-ldap")
+    assert 'experimental support for draft-vchu-ldap-pwd-policy' in log['2.4.7']
+    assert 'defines for SASL and SSL in setup.cfg' in log['2.4.4']
+    assert "'strf_secs' and 'strp_secs' to ldap.functions" in log['2.4.26']
+
+
 def test_pyinotify():
     log = changelogs.get("pyinotify")
     assert 'IN_Q_OVERFLOW handling' in log['0.9.3']