+++ changelogs/custom/pypi/pyparsing.py
@@ -1,4 +1,0 @@
-def get_urls(releases, **kwargs):
-    return [
-               "https://sourceforge.net/p/pyparsing/code/HEAD/tree/trunk/src/CHANGES?format=raw"
-    ], set()