+++ HISTORY.rst
@@ -6,7 +6,8 @@ History
 --------------------
 
 * Fix issue with custom parsing of packages with different case.
-
+* Added custom parser:
+  - alembic
 
 0.8.0 (2016-03-29)
 ------------------