+++ HISTORY.rst
@@ -15,6 +15,7 @@ History
   - genshi
   - imapclient
   - mako
+  - pyinotify
 
 0.8.0 (2016-03-29)
 ------------------