+++ HISTORY.rst
@@ -18,6 +18,7 @@ History
   - mako
   - pyinotify
   - python-ldap
+  - redis
 
 0.8.0 (2016-03-29)
 ------------------